Refactor tick_engine into package
